bootcamp_backup_2025-08-29_12-51-03/requirements.txt
flask>=3.0
scikit-learn>=1.3
pandas>=2.2
numpy>=1.26
joblib>=1.3
python-dotenv>=1.0